requirements.txt
streamlit==1.32.0
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.26.2
torch==2.1.2
torchvision==0.16.2
clip-interrogator==0.6.0
open-clip-torch==2.23.0
transformers==4.36.2
moviepy==1.0.3 